# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Arbor API
  summary: 多租户人设工作台 HTTP 契约
  version: 0.1.0
  description: |
    入站适配器契约。实现必须把租户与人设边界留在应用层/仓储端口，
    而不是仅在文档里声明。
servers:
  - url: /v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
    PersonaId:
      name: persona_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ThreadId:
      name: thread_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              description: 每次错误单独生成的 ULID
...
    Citation:
      type: object
      properties:
        memory_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        preview:
          type: string
    AssistantMessage:
      type: object
      properties:
        message_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [assistant]
        text:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        inbox_created:
          type: integer
        attachments:
          type: array
          items:
            type: object
            properties:
              filename:
                type: string
...
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - bearerAuth: []
paths:
...
  /personas/{persona_id}/threads:
    get:
      tags: [Threads]
      summary: 列出会话
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
      responses:
        "200":
          description: 会话列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        updated_at:
                          type: string
                          format: date-time
    post:
      tags: [Threads]
      summary: 创建会话（一人设一会话绑定）
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  persona_id:
                    type: string
                    format: uuid
        "403":
          $ref: "#/components/responses/Error"
  /threads/{thread_id}/messages:
    get:
      tags: [Threads]
      summary: 消息分页
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/ThreadId"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: 消息
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        role:
                          type: string
                        content:
                          type: string
                        citations:
                          type: array
                          items:
                            type: string
                        attachments:
                          type: array
                          items:
                            type: object
                            properties:
                              filename:
                                type: string
                  total:
                    type: integer
                    description: 分页前的条数
        "404":
          $ref: "#/components/responses/Error"
    post:
      tags: [Threads]
      summary: 发送消息并生成带引用的回复
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/ThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                attachments:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  type: string
                file:
                  type: string
                  format: binary
                  description: 默认上限 32MB
      responses:
        "200":
          description: 助手消息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantMessage"
        "403":
          $ref: "#/components/responses/Error"
  /threads/{thread_id}/attachments/{filename}:
    get:
      tags: [Threads]
      summary: 下载会话附件
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/ThreadId"
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 文件字节
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/Error"
  /threads/{thread_id}/export:
    post:
      tags: [Threads]
      summary: 导出会话 JSON 并记审计
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/ThreadId"
      responses:
        "200":
          description: 会话快照
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  persona_id:
                    type: string
                    format: uuid
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        role:
                          type: string
                        content:
                          type: string
                        citations:
                          type: array
                          items:
                            type: string
        "404":
          $ref: "#/components/responses/Error"
...